# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvitationAdmissionOut": {
        "properties": {
          "notification_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Id"
          },
          "queued": {
            "title": "Queued",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "notification_id",
          "queued",
          "reason"
        ],
        "title": "InvitationAdmissionOut",
        "type": "object"
      },
      "InvitationPageOut": {
        "properties": {
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "not_requested",
                "requires_assistance",
                "active",
                "pending",
                "sent",
                "failed",
                "expired",
                "activated",
                "superseded"
              ]
            },
            "title": "Counts",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvitationStatusOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "counts",
          "total",
          "next_cursor"
        ],
        "title": "InvitationPageOut",
        "type": "object"
      },
      "InvitationRequestIn": {
        "properties": {
          "resend": {
            "default": false,
            "title": "Resend",
            "type": "boolean"
          },
          "user_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "User Ids",
            "type": "array"
          }
        },
        "required": [
          "user_ids"
        ],
        "title": "InvitationRequestIn",
        "type": "object"
      },
      "InvitationStatusOut": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "notification_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "state": {
            "enum": [
              "not_requested",
              "requires_assistance",
              "active",
              "pending",
              "sent",
              "failed",
              "expired",
              "activated",
              "superseded"
            ],
            "title": "State",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "full_name",
          "role",
          "notification_id",
          "state",
          "reason",
          "created_at",
          "sent_at",
          "expires_at"
        ],
        "title": "InvitationStatusOut",
        "type": "object"
      },
      "InvitationsQueuedOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/InvitationAdmissionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "notification_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Notification Ids",
            "type": "array"
          },
          "queued": {
            "title": "Queued",
            "type": "integer"
          },
          "skipped": {
            "title": "Skipped",
            "type": "integer"
          }
        },
        "required": [
          "queued",
          "skipped",
          "notification_ids",
          "items"
        ],
        "title": "InvitationsQueuedOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "__Host-nalar_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "NALAR Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/schools/{school_id}/account-invitations": {
      "get": {
        "operationId": "list_invitations_api_v1_schools__school_id__account_invitations_get",
        "parameters": [
          {
            "in": "path",
            "name": "school_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "School Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "not_requested",
                    "requires_assistance",
                    "active",
                    "pending",
                    "sent",
                    "failed",
                    "expired",
                    "activated",
                    "superseded"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationPageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Invitations",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "request_invitations_api_v1_schools__school_id__account_invitations_post",
        "parameters": [
          {
            "in": "path",
            "name": "school_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "School Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationsQueuedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Request Invitations",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}